+++ src/BayesFilters/include/BayesFilters/GaussianPrediction.h
@@ -29,7 +29,7 @@ public:
 
     bool skip(const std::string& what_step, const bool status) override;
 
-    bool getSkipState() override;
+    bool is_skipping() override;
 
     virtual StateModel& getStateModel() noexcept = 0;