+++ src/BayesFilters/include/BayesFilters/StateModel.h
@@ -26,7 +26,7 @@ public:
 
     bool skip(const std::string& what_step, const bool status) override;
 
-    bool getSkipState() override;
+    bool is_skipping() override;
 
     bool add_exogenous_model(std::unique_ptr<ExogenousModel> exogenous_model);